requirements.txt
flair==0.13.0
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
cachetools==5.3.2
transformers==4.30.2
python-dotenv==1.0.1
openai==1.14.0